third_party/blink/web_tests/svg/custom/svg-viewBox-dynamic-expected.txt

Tests whether removing the viewBox attribute propagates to the viewBox DOM values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS svg.viewBox.baseVal.x is 20
PASS svg.viewBox.baseVal.y is 10
PASS svg.viewBox.baseVal.width is 200
PASS svg.viewBox.baseVal.height is 100
PASS svg.viewBox.baseVal.x is 0
PASS svg.viewBox.baseVal.y is 0
PASS svg.viewBox.baseVal.width is 0
PASS svg.viewBox.baseVal.height is 0
PASS svg.viewBox.baseVal.x is 20
PASS svg.viewBox.baseVal.y is 10
PASS svg.viewBox.baseVal.width is 200
PASS svg.viewBox.baseVal.height is 100
PASS svg.viewBox.baseVal.x is 0
PASS svg.viewBox.baseVal.y is 0
PASS svg.viewBox.baseVal.width is 0
PASS svg.viewBox.baseVal.height is 0
PASS svg.getAttribute('viewBox') is 'invalid value'
PASS successfullyParsed is true

TEST COMPLETE
